+++ tapo-py/examples/tapo_p300.py
@@ -1,4 +1,4 @@
-"""P300 and P304M Example"""
+"""P300, P304M and P316M Example"""
 
 import asyncio
 import os