+++ tapo/src/error.rs
@@ -1,25 +1,34 @@
 /// Response Error from the Tapo API.
-#[derive(Debug)]
+#[derive(thiserror::Error, Debug)]
 #[non_exhaustive]
 pub enum TapoResponseError {
     /// Invalid request.
+    #[error("Invalid request")]
     InvalidRequest,
     /// Invalid response.
+    #[error("Invalid response")]
     InvalidResponse,
     /// Malformed request.
+    #[error("Malformed request")]
     MalformedRequest,
     /// Parameters were invalid
+    #[error("Invalid parameters")]
     InvalidParameters,
     /// Invalid public key.
+    #[error("Invalid public key")]
     InvalidPublicKey,
     /// The credentials provided were invalid.
-    InvalidCredentials,
+    #[error("{0}")]
+    InvalidCredentials(String),
     /// Session timeout.
+    #[error("Session timeout")]
     SessionTimeout,
     /// Unexpected empty result.
+    #[error("Unexpected empty result")]
     EmptyResult,
     /// Unknown Error. This is a catch-all for errors that don't fit into the other categories.
     /// In time, some of these might be added as their own variants.
+    #[error("Unknown error: {0}")]
     Unknown(i32),
 }
 
@@ -28,7 +37,7 @@ pub enum TapoResponseError {
 #[non_exhaustive]
 pub enum Error {
     /// Response Error from the Tapo API.
-    #[error("Tapo: {0:?}")]
+    #[error(transparent)]
     Tapo(TapoResponseError),
     /// Validation Error of a provided field.
     #[error("Validation: {field} {message}")]
@@ -39,10 +48,10 @@ pub enum Error {
         message: String,
     },
     /// Serialization/Deserialization Error.
-    #[error("Serde: {0}")]
+    #[error(transparent)]
     Serde(#[from] serde_json::Error),
     /// HTTP Error.
-    #[error("Http: {0}")]
+    #[error(transparent)]
     Http(#[from] reqwest::Error),
     /// Device not found
     #[error("Device not found")]